# docs/swagger.yaml
definitions:
  dopTypes.CreateRep:
    properties:
      id: {}
    type: object
  dopTypes.ErrRep:
    properties:
      desc:
        type: string
      error_code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
  dopTypes.PaginatedListRep:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      results: {}
      total_count:
        type: integer
    type: object
  entities.AppCUSt:
    properties:
      name:
        type: string
      perm_url:
        type: string
    type: object
  entities.AppSt:
    properties:
      id:
        type: integer
      is_account_app:
        type: boolean
      name:
        type: string
      perm_url:
        type: string
    type: object
  entities.AuthByTokenRepSt:
    properties:
      access_token:
        type: string
    type: object
  entities.AuthByTokenReqSt:
    properties:
      refresh_token:
        type: string
    type: object
  entities.AuthRepSt:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  entities.ConfigSt:
    properties:
      access_token_dur_seconds:
        type: integer
      refresh_token_dur_seconds:
        type: integer
    type: object
...
  entities.PermCUSt:
    properties:
      app_id:
        type: integer
      code:
        type: string
      dsc:
        type: string
      is_all:
        type: boolean
    type: object
  entities.PermSt:
    properties:
      app_id:
        type: integer
      code:
        type: string
      dsc:
        type: string
      id:
        type: integer
      is_all:
        type: boolean
      is_system:
        type: boolean
    type: object
  entities.PhoneAndSmsCodeSt:
    properties:
      phone:
        type: string
      sms_code:
        type: integer
    type: object
...
  entities.RoleListSt:
    properties:
      code:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
    type: object
...
  entities.SendPhoneValidatingCodeReqSt:
    properties:
      err_ne:
        type: boolean
      phone:
        type: string
    type: object
  entities.UsrCUSt:
    properties:
      ava:
        type: string
      name:
        type: string
      phone:
        type: string
      role_ids:
        items:
          type: integer
        type: array
    type: object
...
  entities.UsrProfileSt:
    properties:
      ava:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      perm_codes:
        items:
          type: string
        type: array
      perm_ids:
        items:
          type: integer
        type: array
      perms:
        items:
          $ref: '#/definitions/entities.PermSt'
        type: array
      phone:
        type: string
      role_codes:
        items:
          type: string
        type: array
      role_ids:
        items:
          type: integer
        type: array
      roles:
        items:
          $ref: '#/definitions/entities.RoleListSt'
        type: array
    type: object
  entities.UsrRegReqSt:
    properties:
      ava:
        type: string
      name:
        type: string
      phone:
        type: string
      sms_code:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /app:
    get:
      parameters:
      - in: query
        items:
          type: string
        name: cols
        type: array
      - in: query
        items:
          type: integer
        name: ids
        type: array
      - in: query
        name: name
        type: string
      - in: query
        name: only_count
        type: boolean
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        items:
          type: string
        name: sort
        type: array
      - in: query
        name: sort_name
        type: string
      - in: query
        name: with_total_count
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dopTypes.PaginatedListRep'
            - properties:
                results:
                  items:
                    $ref: '#/definitions/entities.AppSt'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - app
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.AppCUSt'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dopTypes.CreateRep'
            - properties:
                id:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - app
  /app/:id:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - app
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AppSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - app
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.AppCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - app
  /config:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ConfigSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      summary: Get configs
      tags:
      - config
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.ConfigSt'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      summary: Update configs
      tags:
      - config
...
  /perm:
    get:
      parameters:
      - in: query
        name: app_id
        type: integer
      - in: query
        name: code
        type: string
      - in: query
        items:
          type: string
        name: cols
        type: array
      - in: query
        items:
          type: integer
        name: ids
        type: array
      - in: query
        name: is_system
        type: boolean
      - in: query
        name: only_count
        type: boolean
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        items:
          type: string
        name: sort
        type: array
      - in: query
        name: sort_name
        type: string
      - in: query
        name: with_total_count
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.PermSt'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - perm
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.PermCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dopTypes.CreateRep'
            - properties:
                id:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - perm
  /perm/:id:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - perm
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.PermSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - perm
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.PermCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - perm
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.UsrProfileSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
    put:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.UsrCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
  /profile/auth:
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.PhoneAndSmsCodeSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
  /profile/auth/token:
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.AuthByTokenReqSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AuthByTokenRepSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
  /profile/change_phone:
    put:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.PhoneAndSmsCodeSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
  /profile/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
  /profile/reg:
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.UsrRegReqSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AuthRepSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
  /profile/send_validating_code:
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.SendPhoneValidatingCodeReqSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - profile
...
swagger: "2.0"